security/plugins: add headless-safe command policy tests; gate flaky scheduler test in headless CI; stabilize UI buffer snapshots; add plugin-loader host_execute_command policy tests; minor logging import fix

+++ openagent-terminal/tests/ui_buffer_snapshots.rs
@@ -51,7 +51,7 @@ fn ui_buffer_wrapping_snapshot() {
         if col >= grid.columns() {
             col = 0;
             row += 1;
-            if row >= grid.screen_lines() { break; }
+            if (row as usize) >= grid.screen_lines() { break; }
         }
     }
